refact: extract required items to its own component
