+++ frontend/components/ui/BlueprintRequiredItems/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./blueprint-required-items.component"
